# Fix escape() to leverage Cow for performance