@djc 2018-11-05
I think this will be obviated by my new branch: https://github.com/djc/askama/tree/escaping.